Fix confusing err message from `assert_not_equals`
The `assert_not_equals` function printed the following error message
when the values were equal:

    Received unexpected value: <value>

This message sounds quite confusing because in this case the question
is raised: 'what is the expected value then?'. But the expected value
is any value different from <value>. So the error message was changed
to

    Actual and expected values are equal: <value>

which is more comprehensible.

Closes #206