Check hooks are defined only for existing tests
When a named test hook, that doesn't belong to any test (for example, if
someone makes a typo in test name), was being defined, it was impossible
to know about that. This might lead to unnecessary debugging actions
from the developer's side to understand what's going on and why the hook
is not running.

Now to catch such cases luatest checks that hooks are defined only for
existing tests and raises an error with a message like that:

    There is no test with name '<test name>' but hook 'before_test' is
    defined for it

which should give glue to developers what is wrong.

Closes #205