Fix confusing doc strings and err messages
This patch fixes confusing docstrings for `assert_items_equals` and
`assert_items_include` functions and error messages that the functions
produce when asserting fails.

These functions do a comparison of actual and expected tables only by
item values (keys are not considered) but docstrings and error messages
made people think that item keys are considered as well which is wrong.

Closes #117