Raise error if net_box_uri > max unix socket path
Since luatest supports in luatest/server.lua the socket connection to
tarantool instances via the net_box_uri param, we need to check that
the socket's path length is less than the max length of the Unix domain
socket path allowed by a system. Otherwise, we will get the `No buffer
space available` error which is not clear for users and may take some
time to understand what's wrong.

Closes #152