Describe positional arguments in the help message
According to the README, CLI arguments are treated as files or dir paths
when an argument contains '/' in it.

The help message confuses users who want to pass a test
filename to luatest:

    Usage: luatest [options] [files or dirs ...] [testname1 ...]

To be consistent with the README, the Usage section is changed to:

    Usage: luatest [options] [path ...] [group ...]

`testname1` changed with `group` to be consistent, as well as the
description which is moved from the Options section to the new
Positional arguments section.

Closes #155