Add new module replica_proxy.lua
This patch adds a new module replica_proxy.lua to the repo that is
a copy of the test/luatest_helpers/proxy/proxy.lua module [1] in the
tarantool/tarantool repo. The later should be deleted in the nearest
future.

The new module is intended to manage connection to Tarantool replication
instances via proxy.

Also, the replica_conn.lua module is added that is considered as an
internal one and not exposed to public API.

[1] tarantool/tarantool@ad42084/test/luatest_helpers/proxy/proxy.lua

Part of #251