Add new module server_v2.lua
This patch adds a new module server_v2.lua to the repo that is an
almost exact copy of the test/luatest_helpers/server.lua module [1]
in the tarantool/tarantool repo. The later should be deleted in the
nearest future.

The new module should be considered as an enhanced version of the old
one, where some significant features like waiting until the server is
up are available.

Also, the server_v2_instance.lua file is added that is considered as
an internal one and not exposed to public API.

[1] tarantool/tarantool@bab289c/test/luatest_helpers/server.lua

Part of #239