Merge server_v2.lua into server.lua
It was decided to merge module server_v2.lua into server.lua to have one
common interface for managing Tarantool instances. The new server.lua is
fully compatible with the old version, which means that nothing changed
for the projects using luatest previously.

In the same time, the new server.lua has one significant feature that
differs it from the old version: by default, it waits until a server is
ready if no custom executable path to start the server is specified.
Otherwise, we fall back to the old behavior.

Part of #239