Fix issue with server start when luatest installed as rock
When luatest is installed as a rock, the internal server_instance.lua
script won't have execution permissions even though it has them in the
source tree, and won't be able to be run while a server start.

Let's start a server process as

    /abs/path/to/tarantool /abs/path/to/luatest/server_instance.lua

instead of just

    /abs/path/to/luatest/server_instance.lua

to bypass the problem.

Also, the shebang string and x-bit are removed from server_instance.lua.

Part of #269