Improve luatest auto-requiring in `Server:exec`
In the commit tarantool/luatest@58eae75 auto-requiring of the 'luatest'
lib was added to the `Server:eval()` function, so it was also available
in `Server:exec()`, and it was defined as `t` variable. In the process
of development we found that the following construction doesn't work due
to the upvalue error:

    -- foo_bar_test.lua
    local t = require('luatest')

    ...
    g.test_foo_bar = function()
        server:exec(function()
            t.assert_equals(1, 1)  -- upvalues exception (!)
        end)
    end

To fix this error we decided to make the global variable `t` that had
the 'luatest' lib already loaded. So there was no need to import the
'luatest' lib in the test files and the following example worked:

    -- foo_bar_test.lua
    -- local t = require('luatest')  -- `t` is already defined

    ...
    g.test_foo_bar = function()
        server:exec(function()
            t.assert_equals(1, 1)  -- it works
        end)
    end:

But such an approach is not explicit enough and may mislead developers.
So it was decided to improve and extend the existing mechanism. Now we
just make the needed upvalue with the 'luatest' lib inside available in
the `Server:exec()` function via some magic. And now we can do something
like this:

    -- foo_bar_test.lua
    local t1 = require('luatest')
    local t2 = require('luatest')

    ...
    g.test_foo_bar_1 = function()
        server:exec(function()
            t1.assert_equals(1, 1)  -- it works
        end)
    end

    g.test_foo_bar_2 = function()
        server:exec(function()
            t2.assert_equals(1, 1)  -- it works
        end)
    end

Unfortunately, we cannot do the same for the `Server:eval()` function
and now it becomes just a shortcut for `Server.net_box:eval()` as it
was before the first version of auto-requiring of the 'luatest' lib.

Follows up #277
In scope of #233