Fix server socket collisions
Now, by default the server socket file is located in the server working
directory which is unique.

If the server belongs to a replica set, then the server socket path will
be also unique since the server socket file is located in the working
directory of the replica set which is unique as well.

Resolves #297