Fix random test hangs when using proxy module
When one end of the connection is closed for some reason, the proxy
should close the other end as well. Otherwise random hangs are possible
when a client doesn't notice that the server has closed the socket.

Previously clients never noticed leader disconnect (they only noticed
that the heartbeats were missing), and some tests relied on that
behaviour while proxy is paused. To preserve it, do not close the
sockets while proxy is paused.

Resolves #298