Add error handling at the server instance
When executing the following code on the remote server instance using
`server:exec()`, we could get an unobvious error:

        s = Server:new()
        s:start()

        -- we expected a custom error with the "My error" message, but
        -- got "Transaction is active at return from function"
        s:exec(function() box.begin() error("My error") end)

This is due to the specific evaluation and capture of the return values
of `net_box:eval` (for more details see here: tarantool/tarantool#7288).

The error will be raised on the server instance side. In the process of
development we found that our approach does not work in the specific
case:

        -- an error message has been expected but got an error table:
        -- {
        --    class = LuatestError,
        --    status = fail,
        --    message = "a value evaluating to true, actual: false"
        -- }
        s:exec(function() require('luatest').assert(false) end)

To fix this error we decided to add the encoding an error to JSON format
on the server and decode it on the client side.

Close #242

Co-authored-by: Yaroslav Lobankov <y.lobankov@tarantool.org>