Restrict proxy module usage to Tarantool 2.10.1+
The proxy module was initially introduced as a helper in tarantool repo
in commit tarantool/tarantool@b907e7260920. This happened around
Tarantool version 2.11, and the module was never used for older
Tarantool versions.

It was discovered that the module fails on Tarantool version 2.10.0-28
and before due to the problem fixed by tarantool/tarantool@57ecb6cd90b4.

Let's restrict proxy module usage to Tarantool 2.10.1 and higher.

Part-of #298