Fix wait_for_downstream_to
Currently `wait_for_downstream_to` doesn't wait for replica to be up
to date with master. It only checks, that downstream "knows" correct
vclock of the replica by taking vclock of remote server and waiting
for info.downstream.vclock to be the same. The function's comment says
the opposite, so this is not expected behavior.

Let's take local vclock instead. This way the function will wait for
remote server to apply all master's data. Moreover, let's update
function's comment to be less vague.

Closes tarantool/tarantool-qa#311