Not require own test helper when running tests
> Fow what test/helper.lua is needed?

Luatest automatically requires `test/helper.lua` file if it's present.
Usually, it is used to configure luatest or run any bootstrap code.

> What's the problem?

To run tarantool functional tests we use the test-run framework [1].
test-run has an integration with luatest (uses luatest as a submodule)
and allows us to run tarantool functional tests written in the luatest
approach. Most of the such tests can be run via luatest directly, but
test-run is used to run all types of tarantool tests at the same time
(diff, python, app, luatest).

When test-run works, it tunes some environment variables for own needs
and sets LUA_PATH to something like this:

    LUA_PATH='<...>/tarantool/?/init.lua;<...>/tarantool/?.lua;\
        <...>/tarantool/test-run/lib/luatest/?/init.lua;\
        <...>/tarantool/test-run/lib/luatest/?.lua;;'

The tarantool tests don't have own `test/helper.lua` file and this file
is required from the luatest submodule of test-run in accordance with
the configured LUA_PATH variable, which is obviously wrong.

> What's the solution?

The simplest way to resolve this issue is to move `test/helper.lua` to
`test/helpers/general.lua` to not auto-require it while running tests.

Actually, there is no reason to have `test/helper.lua` in luatest. I can
assume it was added to run unit tests with `shuffle = 'group'`. But we
can run unit tests with the `--shuffle group` option and have the same
effect.

[1] https://github.com/tarantool/test-run

Closes #306