Accept table as new_box_uri
Prior to this patch, box.cfg.listen was only set to string or number,
although table is also supported by require('uri').parse(). Because of
this, it was not possible to use the table as net_box_uri when creating
the server. This patch allows the table to be used as net_box_uri. Note,
that for backward compatiblity net_box_uri given as table is converted
to string before setting as box.cfg.listen. This is done because most of
server implementations support box.cfg.listen as string.

Closes #342