Change conversion of `box.error` to string in error message assertions
Currently, error message assertions convert `box.error` to string using
`tostring`. The string conversion metamethod of `box.error` is currently
equivalent to the `message` field of `box.error`. In scope of
tarantool/tarantool#9105, we are going to increase the verbosity of
the string metamethod, and it will no longer be suitable for testing error
messages. Let's replace `tostring(error)` conversions with `error.message`
to retain the old behaviour.

Needed for tarantool/tarantool#9105