server: convert `box.error` thrown from net.box's eval to string
net.box's `eval` method always throws a `box.error` in case of an error.
When we decode the `eval` we result, in case of an error we currently rely
on the fact that the string conversion metamethod returns the error's
message.

However, in scope of tarantool/tarantool#9105, we are going to increase the
verbosity of the string conversion metamethod, and it will no longer return
the error's message. Let's explicitly extract the error's message to retain
the old behaviour.

Needed for tarantool/tarantool#9105