Add --list-test-cases and --run-test-case
These options allow an external runner (such as [test-run][test-run]) to
run the test cases separately: isolated and/or parallel.

The idea is the following.

* test-run has some criteria, whether to parallelize the given test.
* If it needs to be parallelized, test-run invokes `luatest
  --list-test-cases <..path to test..>`.
* Then each test case is run in its own process using `luatest
  --run-test-case <..test case..> <..path to test..>`.

The `--run-test-case` option may be simulated using `--pattern`, but it
requires regexp escaping. It is more convenient to have a separate
option to pass a verbatim test case name.

The new CLI options are mostly for scripting and not for interactive
use. So, no short option are added for them.

[test-run] https://github.com/tarantool/test-run