Do not spam stdout on failure when unified logging is enabled
Unified logging is a fetaure that enables writing logs emitted by all
server instances instances and the luatest runner itself to a single
file. It's implemented by duplicating logs to stdout (using the tee
Linux command) and piping stdout to a log file in OutputBeautifier.
The OutputBeautifier still writes the logs to stdout, even if unified
logging is enabled. As a result, on any failure the runner spams stdout
with captured logs of started server instances. There's no need for
that because all logs can be read from the unified log file. It only
clutters the output making it difficult to figure out what went wrong.
So let's disable writing to stdout in OutputBeautifier if unified
logging is enabled.

No changelog because the feature was added by commit f8a1c10baa482
("Add logging to unified file"), which hasn't been released yet.

Closes #382