Do not pretty format log arguments
This is bad because it quotes all strings although most of the times we
don't need it. This also makes `luatest.log` work differently from the
built-in logger while we want them to be interchangeable.

We don't need pretty-printing anymore because we don't log assertion
arguments. If the caller needs pretty-printing, they can do it manually.

This effectively reverts commit 15dbf758473f0 ("Improve luatest.log
function if a nil value is passed") and a good part of commit
a005329785a1d ("Add syntax sugar for more convenient logging").