Print luatest logs to stdout when unified log is disabled
If unified logging is enabled, logs written by the runner are written
both to its own log file and to the unified log file. This is consistent
with how logs written by test processes are handled. However, if unified
logging is disabled, then runner logs are written only to its log file
while test process logs are also written to stdout and captured if
capturing is enabled. Let's write runner logs to stdout in this case as
well because they can be useful for debugging.

While we are at it, let's get rid of using `tee` for logging: we can
redirect all logs to the output beautifier directly and instruct the
latter to write them to a file.

No changelog because the feature was added by commit f8a1c10baa482
("Add logging to unified file"), which hasn't been released yet.