server: fix grep_log not finding string logged immediately before
Logs written by a test server go through a pipe and are processed by
a luatest fiber before they make it to the log file so if grep_log()
is called immediately after exec() that writes something to the log
it may fail to find the logged string. Let's fix this issue by writing
a unique marker string to the server log via exec() and retrying
grep_log() until we find the marker.

Closes #421