Make `assert_covers` check array items for coverage
Treating arrays as maps in `assert_covers` is error-prone so we fixed it
in commit 95a1c3facc4a ("Fix `assert_covers` treat arrays as maps"), but
we did it over-zealously - now arrays are treated just like scalars
(numbers, for example). This breaks the case when array items are
supposed to be checked for coverage for example:

```lua
t.assert_covers({
    {a = 1, b = 10},
    {a = 2, b = 20},
    {a = 3, b = 30},
}, {{a = 1}, {a = 2}, {a = 3}})
```

Fix this.

Follow-up #405
Closes #439